+++ api/index.ts
@@ -2,27 +2,28 @@ import express from "express"
 import cors from "cors"
 import corsOptions from "./config/corsOptions"
 import mysql from "mysql2/promise"
+import dotenv from "dotenv"
 
-import dbMiddleware from "./config/dbPool"
+import dbMiddleware from "./middlewares/dbMiddleware"
 
 import loginRoute from "./users/login"
 import registerRoute from "./users/register"
+
+dotenv.config()
 
 const app = express()
 
 app.use(cors(corsOptions))
 app.use(express.json())
 
-const pool = mysql.createPool({
-    host: process.env.DB_HOST,
-    user: process.env.DB_USER,
-    password: process.env.DB_PASSWORD,
-    database: process.env.DB_DATABASE,
-})
-
-app.use(dbMiddleware(pool))
+app.use(dbMiddleware())
 
 app.use("/api/users", loginRoute)
 app.use("/api/users", registerRoute)
+
+const PORT = process.env.PORT || 8000
+app.listen(PORT, () => {
+    console.log(`Server is running on port ${PORT}`)
+})
 
 export default app
