+++ nodemon.json
@@ -1,0 +1,6 @@
+{
+    "watch": ["src"],
+    "ext": "ts",
+    "ignore": ["node_modules"],
+    "exec": "ts-node ./src/index.ts"
+}
